@jkotas 2018-09-14
You can every other method in the framework, mark with AggressivelyInlining, and show that that it is improving perf using microbenchmark. I do not think we should be doing that - the microbenchmark are not capturing the code bloat that this is producing and that will slow down real world scenarions.