@jkotas 2018-09-14
> I will wait for ML.NET to consume this API and measure

This is a good plan. If you can demonstrate that the applying AggressiveInlining here materially improves ML.NET performance, we can reconsider.

I see ML.NET using the Trim methods in pretty big and complex methods (for example:   https://github.com/dotnet/machinelearning/pull/863/files#diff-4f49b4fddbc210a67cbeb1678bbaf21cR775). Big methods like this will suffer from register pressure. I can bet money on that force-inlining more non-trivial methods like `TrimWhiteSpace` into already big methods is going to regress the performance, not improve it.