@adamsitnik 2018-09-17
> @adamsitnik, can we do something about how string data shows up in the BDN output for GitHub so its clearly visible, especially if it contains white space? Surrounding it with quotes might help (I manually formatted the output in the "Before" case above).

@ahsonkhan Sure, I would be happy to approve such PR ;)