@jkotas 2018-09-20
> Trim isn't marked as AI, large/complex methods calling Trim shouldn't be negatively affected.

Trim is trivial method, so the JIT will be likely to inline it most of the time. Once it inlines it, it will see the TrimStart/End methods marked with AggresiveInlining and it will try hard to inline those as well because of AggresiveInlining flag told it to do so.

If you just want to make the non-inlined Trim better, you should change it to streamlined implementation. Something like this:

```csharp
ReadOnlySpan<char> Trim(this ReadOnlySpan<char> span, char trimChar)
{
    int start = 0;
    for (; start < span.Length; start++)
    {
        if (span[start] != trimChar)
            break;
    }
    int end = span.Length - 1;
    for (; end >= start; end--)
    {
        if (span[end] != trimChar)
            break;
    }
    return span.Slice(start, end - start + 1);
}
```

It will avoid the extra call and it will also avoid other minor inefficiencies from passing the arguments around and calling Slice twice instead of just once.