@ahsonkhan 2018-09-20
> Trim is trivial method, so the JIT will be likely to inline it most of the time. 

Good point.

> If you just want to make the non-inlined Trim better, you should change it to streamlined implementation. 

Made this change.

Thanks for the feedback.